Add fft grid and raz grid to test against master
