+++ tests/test_compute_everything.py
@@ -24,13 +24,14 @@ from desc.geometry import (
     FourierXYZCurve,
     ZernikeRZToroidalSection,
 )
-from desc.grid import LinearGrid
+from desc.grid import Grid, LinearGrid
+from desc.integrals import Bounce2D
 from desc.magnetic_fields import (
     CurrentPotentialField,
     FourierCurrentPotentialField,
     OmnigenousField,
 )
-from desc.utils import ResolutionWarning, errorif, xyz2rpz, xyz2rpz_vec
+from desc.utils import ResolutionWarning, apply, errorif, xyz2rpz, xyz2rpz_vec
 
 
 def _compare_against_master(
@@ -259,6 +260,11 @@ def test_compute_everything():
                 f"Parameterization: {p}. Can't compute "
                 + f"{names - this_branch_data_rpz[p].keys()}."
             )
+
+            # now we get the special grid data
+            this_branch_data_rpz[p].update(fft_grid_data(p))
+            this_branch_data_rpz[p].update(raz_grid_data(p))
+
             # compare data against master branch
             error_rpz, update_master_data_rpz = _compare_against_master(
                 p,
@@ -297,3 +303,59 @@ def test_compute_everything():
             pickle.dump(this_branch_data_rpz, file)
 
     assert not error_rpz
+
+
+def fft_grid_data(p):
+    """Compute fft grid quantities."""
+    if p != "desc.equilibrium.equilibrium.Equilibrium":
+        return {}
+
+    # TODO: can automate this later to add omngeneity, boozer transform, etc.
+    fft_names = ["effective ripple", "Gamma_c", "Gamma_c Velasco"]
+
+    eq = get("W7-X")
+    rho = np.linspace(0, 1, 10)
+    grid = LinearGrid(rho=rho, M=eq.M_grid, N=eq.N_grid, NFP=eq.NFP, sym=False)
+
+    kwargs = dict(
+        angle=Bounce2D.angle(eq, X=32, Y=32, rho=rho),
+        Y_B=grid.num_zeta * grid.NFP,
+        num_transit=10,
+        num_well=20 * 10,
+        nufft_eps=1e-8,
+    )
+    data = eq.compute(fft_names, grid, surf_batch_size=1, **kwargs)
+
+    # check vectorization too
+    d = data.copy()
+    for name in fft_names:
+        d.pop(name)
+    d = eq.compute(fft_names, grid, surf_batch_size=2, data=d, **kwargs)
+    for name in fft_names:
+        np.testing.assert_allclose(
+            d[name], data[name], rtol=1e-8, atol=1e-8, err_msg=name
+        )
+
+    data = apply(data, grid.compress, fft_names)
+    return data
+
+
+def raz_grid_data(p):
+    """Compute field line grid quantities."""
+    if p != "desc.equilibrium.equilibrium.Equilibrium":
+        return {}
+
+    # TODO: can automate this later to add ballooning etc.
+    raz_names = ["old effective ripple", "old Gamma_c", "old Gamma_c Velasco"]
+
+    eq = get("W7-X")
+    num_transit = 10
+    Y_B = eq.N_grid * 2 * eq.NFP
+    rho = np.linspace(0, 1, 10)
+    alpha = np.array([0])
+    zeta = np.linspace(0, num_transit * 2 * np.pi, num_transit * Y_B)
+    grid = Grid.create_meshgrid([rho, alpha, zeta], coordinates="raz")
+
+    data = eq.compute(raz_names, grid, num_well=20 * num_transit)
+    data = apply(data, grid.compress, raz_names)
+    return data
